registry: move event/subscriber plumbing and color picking to sibling files
Pure cut-paste out of registry.go into events.go and colors.go,
following the existing projects.go / persist.go pattern. No behavior
change; no test edits.
